Remove Poetry Lock file
Since this is a package, we have no need of the poetry.lock file since
it is only used during local development, and not when someone installs
this package. Therefore we shouldn't use lock files in libraries, only
on projects in order to have reproducible builds.

More information on using Poetry for packages:
- https://python-poetry.org/docs/libraries/