Allow external code execution during pip updates check
Package managers with the package-ecosystem values such as pip
may execute external code in the manifest as part of the version
update process. This might allow a compromised package to steal
credentials or gain access to configured registries.

Dependabot automatically prevents this behavior and stop updates
during the checking process. We need to override this setting in order
to get pip updates.